+++ src/main/kotlin/com/pylons/wallet/core/engine/TxPylonsDevEngine.kt
@@ -9,7 +9,7 @@ internal class TxPylonsDevEngine : TxPylonsEngine () {
     override val isDevEngine: Boolean = true
     override val backendType: Backend = Backend.LIVE_DEV
 
-    override fun createCookbook(name: String, devel: String, desc: String, version: String, supportEmail: String, level: Int): Transaction =
+    override fun createCookbook(name: String, devel: String, desc: String, version: String, supportEmail: String, level: Long): Transaction =
             basicTxHandlerFlow { TxJson.createCookbook(name, devel, desc, version, supportEmail, level, it.address,
                     cryptoHandler.keyPair!!.publicKey(), it.accountNumber, it.sequence) }
 
@@ -17,11 +17,11 @@ internal class TxPylonsDevEngine : TxPylonsEngine () {
             basicTxHandlerFlow { TxJson.updateCookbook(id, devel, desc, version, supportEmail, it.address,
                     cryptoHandler.keyPair!!.publicKey(), it.accountNumber, it.sequence) }
 
-    override fun createRecipe(name : String, cookbookName: String, desc: String, inputs: Map<String, Int>, outputs: Map<String, Int>, time: Int): Transaction =
+    override fun createRecipe(name : String, cookbookName: String, desc: String, inputs: Map<String, Long>, outputs: Map<String, Long>, time: Long): Transaction =
             basicTxHandlerFlow { TxJson.createRecipe(name, cookbookName, desc, inputs, outputs, time,
                     it.address, cryptoHandler.keyPair!!.publicKey(), it.accountNumber, it.sequence) }
 
-    override fun updateRecipe(name: String, cookbookName: String, id: String, desc: String, inputs: Map<String, Int>, outputs: Map<String, Int>, time: Int): Transaction =
+    override fun updateRecipe(name: String, cookbookName: String, id: String, desc: String, inputs: Map<String, Long>, outputs: Map<String, Long>, time: Long): Transaction =
             basicTxHandlerFlow { TxJson.updateRecipe(name, cookbookName, id, desc, inputs, outputs, time,
                     it.address, cryptoHandler.keyPair!!.publicKey(), it.accountNumber, it.sequence) }
